+++ core/src/exchanges/limitless/index.ts
@@ -139,7 +139,7 @@ export class LimitlessExchange extends PredictionMarketExchange {
         return fetchEvents(params, this.callApi.bind(this), this.http);
     }
 
-    async fetchOHLCV(id: string, params: OHLCVParams | HistoryFilterParams): Promise<PriceCandle[]> {
+    async fetchOHLCV(id: string, params: OHLCVParams): Promise<PriceCandle[]> {
         return fetchOHLCV(id, params, this.callApi.bind(this));
     }