+++ core/src/exchanges/myriad/index.ts
@@ -87,7 +87,7 @@ export class MyriadExchange extends PredictionMarketExchange {
         return fetchEvents(params, this.getHeaders(), this.http);
     }
 
-    async fetchOHLCV(id: string, params: OHLCVParams | HistoryFilterParams): Promise<PriceCandle[]> {
+    async fetchOHLCV(id: string, params: OHLCVParams): Promise<PriceCandle[]> {
         return fetchOHLCV(id, params, this.callApi.bind(this));
     }